requirements.txt
tqdm
google-api-python-client
oauth2client
apiclient
youtube_dl
python-dotenv
beautifulsoup4
youtube_transcript_api
